requirements.txt
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
azure-core==1.29.0
flask==2.3.2
flask-cors==4.0.0
